# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Signal Lab server",
    "version": "1",
    "summary": "The engine of Signal Lab over HTTP: commands, experiment runs, events and files.",
    "description": "The API `signal-lab-server` serves next to its interface. Without a token the server listens only on loopback and needs no authentication; with one, every endpoint except `/api/health` needs `Authorization: Bearer <token>` (scripts) or the session cookie a browser gets at `/login`. Requests that change state must not carry an `Origin` of another site, and take JSON only. Failures are an `EngineError`: a stable `code` (its text is `err.<code>` in the interface's dictionaries), `params`, the `node` and `field` it is about and the system's own wording as `detail`. A guide with examples: https://proanima.github.io/SignalLab/api/.",
    "license": { "name": "MIT", "identifier": "MIT" }
  },
  "servers": [{ "url": "http://127.0.0.1:1430", "description": "A server on this machine (the default address)" }],
  "security": [{ "bearer": [] }, { "session": [] }],
...
  "paths": {
...
    "/api/run": {
      "post": {
        "tags": ["runs"],
        "operationId": "run",
        "summary": "Run an experiment to its end",
        "description": "Starts the same run the Run button starts — a job, listed and stoppable in the interface, its events sent to every page, its report saved under `runs/` in the data folder — and answers when it ends. A run that fails is still `200` with `outcome: failed`; an HTTP error means no run was started.\n\nWith `Accept: application/x-ndjson` the answer is one JSON object per line as the run goes: `started`, a `step` per step, a `heartbeat` every 15 s, then `ended` with the result. Without it the answer is the result alone; until the run ends the body carries a space every 15 s (JSON ignores leading whitespace), so a proxy does not close a quiet connection.\n\nA client that disconnects does not stop the run: it goes on to its end and saves its report. When the server shuts down the run is stopped and the answer ends with `outcome: stopped`.",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RunRequest" } } }
        },
        "responses": {
          "200": {
            "description": "The run ended — passed, failed or stopped",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/RunResult" } },
              "application/x-ndjson": {
                "schema": { "$ref": "#/components/schemas/RunLine" },
                "description": "One `RunLine` per line, `ended` last"
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "415": { "$ref": "#/components/responses/NotJson" },
          "422": {
            "description": "The experiment could not start: it does not validate, a parameter is unknown, a secret is not set on the server, a port it listens on is taken, …",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/EngineError" } } }
          }
        }
      }
    },
...
    "/api/events": {
      "get": {
        "tags": ["server"],
        "operationId": "events",
        "summary": "WebSocket of engine events",
        "description": "Upgrade to a WebSocket (from this server's own origin, or with no Origin and the token). Each message is `{\"event\": <name>, \"payload\": <value>}`: `experiment://step`, `experiment://ended`, `job://ended`, `inspect://batch` and the modules' telemetry. A page that falls behind gets `server://lagged` with how many events it missed. The server pings every 20 s and closes the socket when it stops.",
        "responses": {
          "101": { "description": "Switching to the WebSocket protocol" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
    "/api/files": {
      "get": {
        "tags": ["server"],
        "operationId": "download",
        "summary": "A file the engine wrote — a run report, an export — as a download",
        "description": "Only files inside the server's data folder; `report_path` of a run result is such a path.",
        "parameters": [{ "name": "path", "in": "query", "required": true, "schema": { "type": "string" }, "examples": { "report": { "value": "/data/runs/run-1759400000000-3.json" } } }],
        "responses": {
          "200": { "description": "The file", "content": { "application/octet-stream": { "schema": { "type": "string", "contentMediaType": "application/octet-stream" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "description": "Not a file in the data folder (`file.not_found`)", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/EngineError" } } } },
          "413": { "description": "Larger than 256 MiB (`file.too_large`)", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/EngineError" } } } }
        }
      }
    },
    "/api/openapi.json": {
      "get": {
        "tags": ["server"],
        "operationId": "openapi",
        "summary": "This description",
        "responses": {
          "200": { "description": "OpenAPI 3.1", "content": { "application/json": { "schema": { "type": "object" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearer": { "type": "http", "scheme": "bearer", "description": "The server's token (`--token-file`, `SIGNALLAB_TOKEN`, or the one `--generate-token` keeps in the data folder)" },
      "session": { "type": "apiKey", "in": "cookie", "name": "signallab_session", "description": "Set by signing in at `/login`; HttpOnly, SameSite=Strict" }
    },
    "responses": {
      "BadRequest": { "description": "The request could not be read (`api.run_invalid`, `api.run_source`, `command.args_invalid`)", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/EngineError" } } } },
      "Unauthorized": { "description": "No token or session (`auth.required`)", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/EngineError" } } } },
      "Forbidden": { "description": "A host name the server does not answer to (`auth.host`), or a request from another site (`auth.origin`)", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/EngineError" } } } },
      "NotJson": { "description": "The body is not `application/json` (`command.json_required`)", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/EngineError" } } } }
    },
    "schemas": {
      "EngineError": {
        "type": "object",
        "required": ["code"],
        "properties": {
          "code": { "type": "string", "description": "Stable; the text is `err.<code>`", "examples": ["wait.timeout", "secret.missing", "auth.required"] },
          "params": { "type": "object", "additionalProperties": { "type": "string" }, "description": "The values of the message" },
          "node": { "type": "string", "description": "The id of the node it is about" },
          "field": {
            "type": "object",
            "required": ["key"],
            "properties": { "key": { "type": "string", "description": "The text is `field.<key>`" }, "index": { "type": "integer", "minimum": 1 } }
          },
          "detail": { "type": "string", "description": "The operating system's, parser's or library's own wording" }
        }
      },
      "RunRequest": {
        "type": "object",
        "description": "One experiment: `document` or `template`, not both.",
        "additionalProperties": false,
        "properties": {
          "document": { "type": "object", "description": "An experiment as the app saves and exports it; older versions are migrated" },
          "template": { "type": "string", "enum": ["empty", "http-check", "status-branch", "parallel-flows", "osc-ping-reply", "poll-until-ready", "flaky-api", "fault-phases", "dependency-outage", "websocket-echo"], "description": "A bundled template by file name (`.json` optional)" },
          "overrides": { "type": "object", "additionalProperties": { "type": ["string", "number", "boolean"] }, "description": "Parameter values for this run only; each must be a parameter of the experiment" },
          "profile": { "type": "string", "description": "The profile to run with; an empty string runs with the defaults" },
          "seed": { "type": "integer", "minimum": 0, "maximum": 9007199254740991, "description": "Random values repeat with the same seed" },
          "timeout": { "type": "integer", "minimum": 1, "maximum": 300, "description": "Seconds before the run fails with `run.timeout`; default and maximum 300" }
        },
        "examples": [{ "template": "osc-ping-reply", "overrides": { "device": "10.0.0.20:9000" }, "timeout": 30 }]
      },
      "RunStarted": {
        "type": "object",
        "required": ["job_id", "experiment", "seed", "profile", "overridden", "started_ms"],
        "properties": {
          "job_id": { "type": "integer" },
          "experiment": { "type": "string", "description": "The experiment's name" },
          "seed": { "type": "integer" },
          "profile": { "type": ["string", "null"] },
          "overridden": { "type": "boolean", "description": "Some values came from `overrides`" },
          "started_ms": { "type": "integer", "description": "Milliseconds since the Unix epoch" }
        }
      },
      "RunEvent": {
        "type": "object",
        "description": "One step of a run, as the timeline shows it.",
        "required": ["job_id", "ts", "node_id", "state", "detail", "message_key", "message_params"],
        "properties": {
          "job_id": { "type": "integer" },
          "ts": { "type": "integer", "description": "Milliseconds since the Unix epoch" },
          "node_id": { "type": "string" },
          "state": { "type": "string", "enum": ["running", "passed", "failed", "retry", "repeating", "load"] },
          "detail": { "type": "string", "description": "An English line for logs" },
          "message_key": { "type": ["string", "null"], "description": "The text of what the step did (`exp.step.*`)" },
          "message_params": { "description": "Its values" },
          "vars": { "type": "object", "description": "Variables the step wrote" },
          "error": { "$ref": "#/components/schemas/EngineError" },
          "frame": { "type": "integer", "description": "The Inspector frame a wait matched" },
          "load": { "$ref": "#/components/schemas/LoadMetrics" }
        }
      },
      "LoadMetrics": {
        "type": "object",
        "description": "What an HTTP node's load measured, on its last step (passed or failed).",
        "properties": {
          "planned": { "type": "integer", "description": "Requests the profile adds up to (Poisson: on average)" },
          "sent": { "type": "integer", "description": "Answered or failed" },
          "ok": { "type": "integer", "description": "Answered 2xx" },
          "failed": { "type": "integer" },
          "missed": { "type": "integer", "description": "Due while every worker was busy, and skipped" },
          "duration_ms": { "type": "integer" },
          "rps": { "type": "number", "description": "Requests sent per second" },
          "error_rate": { "type": "number", "description": "Failed, in % of sent" },
          "min_ms": { "type": "number" },
          "mean_ms": { "type": "number" },
          "max_ms": { "type": "number" },
          "p50_ms": { "type": "number" },
          "p90_ms": { "type": "number" },
          "p95_ms": { "type": "number" },
          "p99_ms": { "type": "number" },
          "received_bytes": { "type": "integer" },
          "statuses": { "type": "object", "additionalProperties": { "type": "integer" }, "description": "By status (\"200\") or, without one, by cause (\"timeout\")" },
          "seconds": { "type": "array", "items": { "type": "object", "properties": { "sent": { "type": "integer" }, "failed": { "type": "integer" }, "mean_ms": { "type": "number" } } } },
          "histogram": { "type": "array", "items": { "type": "object", "properties": { "upto_ms": { "type": ["number", "null"] }, "count": { "type": "integer" } } } },
          "thresholds": { "type": "array", "items": { "type": "object", "properties": {
            "metric": { "type": "string", "enum": ["p50_ms", "p90_ms", "p95_ms", "p99_ms", "mean_ms", "max_ms", "error_rate", "rps", "missed"] },
            "op": { "type": "string", "enum": ["lt", "le", "gt", "ge"] },
            "value": { "type": "number" },
            "actual": { "type": "number" },
            "held": { "type": "boolean" } } } }
        }
      },
      "RunResult": {
        "type": "object",
        "required": ["job_id", "experiment", "outcome", "seed", "profile", "overridden", "params", "started_ms", "ended_ms", "steps"],
        "properties": {
          "job_id": { "type": "integer" },
          "experiment": { "type": "string" },
          "outcome": { "type": "string", "enum": ["passed", "failed", "stopped"] },
          "seed": { "type": "integer", "description": "Give it as `seed` to repeat the run's random values" },
          "profile": { "type": ["string", "null"] },
          "overridden": { "type": "boolean" },
          "params": { "type": "object", "additionalProperties": { "type": "string" }, "description": "The parameter values the run used" },
          "started_ms": { "type": "integer" },
          "ended_ms": { "type": "integer" },
          "error": { "$ref": "#/components/schemas/EngineError", "description": "Why it failed (secret values masked)" },
          "steps": { "type": "array", "items": { "$ref": "#/components/schemas/RunEvent" } },
          "report_path": { "type": "string", "description": "The report on the server; download it from `/api/files`. A stopped run has none" },
          "report_error": { "$ref": "#/components/schemas/EngineError" }
        }
      },
      "RunLine": {
        "description": "A line of `application/x-ndjson`.",
        "oneOf": [
          { "allOf": [{ "$ref": "#/components/schemas/RunStarted" }, { "type": "object", "required": ["type"], "properties": { "type": { "const": "started" } } }] },
          { "allOf": [{ "$ref": "#/components/schemas/RunEvent" }, { "type": "object", "required": ["type"], "properties": { "type": { "const": "step" } } }] },
          { "type": "object", "required": ["type"], "properties": { "type": { "const": "heartbeat" } } },
          { "allOf": [{ "$ref": "#/components/schemas/RunResult" }, { "type": "object", "required": ["type"], "properties": { "type": { "const": "ended" } } }] }
        ]
      }
    }
  }
}